requirements.txt
aiohttp_jinja2==1.6
aiohttp_cors==0.7.0
aiohttp==3.11.11
python_pam==2.0.2
PyJWT==2.10.1
pam==0.2.0
